Enum sabiwm::backend::Event

pub enum Event<Window> {
    BackendChanged,
    WindowCreated(Window),
    WindowClosed(Window),
    WindowHid(Window),
    WindowRevealed(Window),
    WindowChangeRequest(Window, Rectangle),
    MouseEnter(Option<Window>),
    MouseLeave(Option<Window>),
    ButtonPressed(Window, Option<Window>),
    ButtonReleased,
    KeyPressed(Window),
    Unknown,
}

A cross-section of all events that can be generated/handled by xlib, xcb and wayland.

Variants
BackendChanged

Something on the backend changed, for example screen got resized or xrandr layout got changed.

WindowCreated(Window)

A new window has been created

WindowClosed(Window)

A window has been closed/killed

WindowHid(Window)

A window has been hid, but is still around

WindowRevealed(Window)

A windowhas been revealed from hidden status

WindowChangeRequest(Window, Rectangle)

A window/app is requesting a change in size

MouseEnter(Option<Window>)

The mouse pointer has entered a window's frame

MouseLeave(Option<Window>)

The mouse pointer has left a window's frame

ButtonPressed(Window, Option<Window>)

A button has been pressed

ButtonReleased

A button has been released

KeyPressed(Window)

A key has been pressed

Unknown

An unknown or not important event