Struct sabiwm::core::Stack

pub struct Stack<T> {
    pub focus: T,
    pub up: Vec<T>,
    pub down: Vec<T>,
}

Handles focus tracking on a workspace. focus keeps track of the focused window's id and up and down are the windows above or below the focus stack respectively.

Immutable

Note that this Stack implementation is immutable and that each operation that would modify it, instead returns a new copy of the Stack with the modified state.

Fields
focus: T up: Vec<T> down: Vec<T>
Methods

impl<T: Debug + Copy + Clone + Eq> Stack<T>

fn new<S: Debug + IntoIterator<Item=T>>(f: T, up: S, down: S) -> Stack<T>

Create a new stack with the given values

Examples

let stack = Stack::new(0, Vec::new(), Vec::new());

Arguments

f - the focussed element to start with up - vector of all elements up the stack down - vector of all elements down the stack

Return value

A new Stack

fn add(&self, t: T) -> Stack<T>

Add a new element to the stack and automatically focus it.

Examples

let stack = Stack::from(42);
let new_stack = stack.add(23);

assert_eq!(2, new_stack.len());

Arguments

t - the new element to add

Return value

The new modified Stack

fn integrate<C: FromIterator<T>>(&self) -> C

Flatten the stack into a new container

Examples

let stack = Stack::new(1, vec![2,3], vec![4,5]);
let v = stack.integrate::<Vec<_>>();

assert_eq!(vec![3,2,1,4,5], v);

Return value

A collection of a flattened Stack

fn filter<F>(&self, f: F) -> Option<Stack<T>> where F: Fn(&T) -> bool

Filter the stack to retain only windows that yield true in the given filter function.

Examples

let stack = Stack::new(1, vec![2,3,4], vec![5,6,7]);
let even_stack = stack.filter(|x| x % 2 == 0).unwrap();

assert_eq!(vec![4,2,6], even_stack.integrate::<Vec<_>>());

Arguments

f - the filter function to use for filtering

Return value

Returns Some(Stack) if there are still items left after filtering or None if there is no element left after filtering.

fn focus_up(&self) -> Stack<T>

Move the focus to the next element in the up list

fn focus_down(&self) -> Stack<T>

Move the focus down

fn swap_up(&self) -> Stack<T>

fn swap_down(&self) -> Stack<T>

fn swap_master(&self) -> Stack<T>

fn reverse(&self) -> Stack<T>

Reverse the stack by exchanging the up and down lists

fn len(&self) -> usize

Return the number of elements tracked by the stack

fn is_empty(&self) -> bool

Checks if the Stack is empty

Examples

let stack = Stack::from(42);
assert_eq!(false, stack.is_empty());

Return value

true if the Stack is empty, false otherwise

fn contains(&self, t: T) -> bool

Checks if the given window is tracked by the stack

let stack = Stack::from(42);
assert_eq!(true, stack.contains(42));
assert_eq!(false, stack.contains(23));

Arguments

t - the element to search for

Return value

true if the Stack contains the given element