Struct sabiwm::core::Workspace

pub struct Workspace<Window> {
    pub id: u32,
    pub tag: String,
    pub stack: Option<Stack<Window>>,
}

Represents a single workspace with a tag (name), id, a layout and a stack for all windows. A workspace is in charge of all windows belonging to that workspace. At each time, a single screen holds one workspace, while all the other workspaces are hidden in the background, while still being managed.

Immutable

Note that this Workspace implementation is immutable and that each operation that would modify it, instead returns a new copy of the Workspace with the modified state.

Fields
id: u32 tag: String stack: Option<Stack<Window>>
Methods

impl<Window: Copy + Clone + PartialEq + Eq + Debug> Workspace<Window>

fn new<S: Into<String>>(id: u32,
tag: S,
stack: Option<Stack<Window>>)
-> Workspace<Window>

Create a new workspace

Examples

let stack = Stack::from(42u32);
let workspace = Workspace::new(0, "Desktop 0", Some(stack));
assert_eq!(1, workspace.len());

fn add(&self, window: Window) -> Workspace<Window>

Add a new window to the workspace by adding it to the stack. If the stack doesn't exist yet, create one.

Examples

let workspace : Workspace<u32> = Workspace::new(0, "Desktop 0", None);
assert_eq!(0, workspace.len());

fn remove(&self, window: Window) -> Workspace<Window>

fn len(&self) -> usize

Returns the number of windows contained in this workspace

fn is_empty(&self) -> bool

fn contains(&self, window: Window) -> bool

Checks if the workspace contains the given window

fn windows(&self) -> Vec<Window>

fn peek(&self) -> Option<Window>

fn map<F>(&self, f: F) -> Workspace<Window> where F: Fn(Stack<Window>) -> Stack<Window>

fn map_option<F>(&self, f: F) -> Workspace<Window> where F: Fn(Stack<Window>) -> Option<Stack<Window>>

fn map_or<F>(&self, default: Stack<Window>, f: F) -> Workspace<Window> where F: Fn(Stack<Window>) -> Stack<Window>